backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
bcrypt
passlib[bcrypt]
python-jose
tensorflow
pandas
openpyxl
pytest
jwt